% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterfunctions.R
\name{cluster.fitted}
\alias{cluster.fitted}
\title{Fuzzy c-medoids clustering on fitted values of splines}
\usage{
cluster.fitted(
  TPSdata,
  center = TRUE,
  addslopes = TRUE,
  k,
  m = 1.2,
  noise = TRUE,
  seed,
  ...
)
}
\arguments{
\item{TPSdata}{Output from the \code{\link[=TPSfit]{TPSfit()}} function.}

\item{center}{A logical expression indicating that trajectories are to be
centered on individual means.}

\item{addslopes}{A logical expression indicating that clustering will be
conducted on the slopes between fitted time-points in addition to the
cross-sectional values at each point.}

\item{k}{An integer value indicating the number of clusters.}

\item{m}{A numeric value called the 'fuzziness parameter.' Must be greater
than 1, with a default of 1.2.}

\item{noise}{A logical expression indicating whether a noise cluster for
outliers is to be included.}

\item{seed}{An optional numeric value for the seed, as the clustering
algorithm uses random start values.}

\item{...}{Additional optional values employed by \pkg{fclust}, including \code{RS},
the number of random starts, and \code{delta}, the distance for outliers.}
}
\value{
An object of class '\code{FKM.TPS}' containing the following components:
\itemize{
\item \code{FKM_TPS }  An object output by the \code{fclust} algorithm.
\item \code{U }   A dataframe containing the degrees of cluster membership for each subject.
\item \code{Umax }  A vector containing the modal cluster assignment for each subject.
\item \code{FKM_TPS_U }  A data frame containing subjects with subject ID, degrees of cluster
membership, and modal class assignment.
\item \code{FKM_indices }  A vector containing the values for six cluster validity indices.
See \code{\link[fclust:Fclust.index]{fclust::Fclust.index()}}
\item \code{wide_data }  A data frame with the values that were clustered on in the
clustering algorithm plus subject IDs.
\item \code{TPSdata }  The '\code{TPSfit}' object that was input into the function.
\item \code{center }  Logical value that was used to determine whether to center on individual
trajectories.
\item \code{noise }  Logical value that was used to determine whether a noise cluster is
included.
\item \code{k }  The number of clusters.
\item \code{m }  The fuzziness parameter.
\item \code{addslopes }  Logical value that was used to determine whether to include the
slopes between fitted time-points.
\item \code{IDmatch }  A data frame matching the original subject ID with new consecutive
ID values.
}
}
\description{
\code{cluster.fitted()} is used to apply a fuzzy c-medoids clustering algorithm to
the fitted values that are output from the \code{TPSfit()} function.
}
\details{
\code{cluster.fitted()} employs package \pkg{fclust} to apply the fuzzy c-medoids
clustering algorithm to the fitted tensor-product spline values that are
output from the \code{\link[=TPSfit]{TPSfit()}} function. The user has options of including a noise
cluster and centering trajectories on the individual means. The \code{addslopes}
option will cluster on the slopes between fitted time-points in addition to
cross-sectional fitting at those points.
}
\examples{
data(TS.sim)

fitsplines <- TPSfit(TS.sim, vars=c("Var1", "Var2", "Var3"), time="Time",
     ID="SubjectID", knots_time=c(0, 91, 182, 273, 365), n_fit_times=10)

clusters1 <- cluster.fitted(fitsplines, k=3, m=1.3, seed=12345, RS=5, noise=TRUE)
summary(clusters1)

}
\seealso{
\itemize{
\item \code{\link[=cluster.coefs]{cluster.coefs()}} is an alternative method that clusters on spline
coefficients rather than fitted values.
\item The \pkg{fclust} R package: \url{https://CRAN.R-project.org/package=fclust}
\item Fuzzy k-medoids function: \code{\link[fclust:FKM.med]{fclust::FKM.med()}}
\item Fuzzy k-medoids function with noise cluster: \code{\link[fclust:FKM.med.noise]{fclust::FKM.med.noise()}}
}
}
