% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterfunctions.R
\name{comp.FKM.indices}
\alias{comp.FKM.indices}
\title{Compare indices for fuzzy clusterings}
\usage{
comp.FKM.indices(clusterings, index = "ALL", clusternames = NA)
}
\arguments{
\item{clusterings}{A list object of the clusterings to compare. Each should
be of class \code{'FKM.TPS'}.}

\item{index}{Desired cluster validity index or indices. Default is \code{"ALL"}.
Options are: \code{"PC"}, \code{"PE"}, \code{"MPC"}, \code{"SIL"}, \code{"SIL.F"}, \code{"XB"}, or
\code{"ALL"}. Multiple options can be input as a character vector. See \code{\link[fclust:Fclust.index]{fclust::Fclust.index()}}}

\item{clusternames}{Optional character vector with names of the clusters being compared.}
}
\value{
A data frame with the desired validity indices for each of the clusterings.
}
\description{
\code{comp.FKM.indices()} compares the clustering indices of two more more
clusterings that are output from \code{\link[=cluster.coefs]{cluster.coefs()}} or \code{\link[=cluster.fitted]{cluster.fitted()}}.
}
\details{
\code{comp.FKM.indices()} compares the clustering indices of two more more
clusterings that are output from \code{\link[=cluster.coefs]{cluster.coefs()}} or \code{\link[=cluster.fitted]{cluster.fitted()}}.
Available indices are those found in the \pkg{fclust} package: \code{PC}
(partition coefficient), \code{PE} (partition entropy), \code{MPC} (modified partition
coefficient), \code{SIL} (silhouette), \code{SIL.F} (fuzzy silhouette), and \code{XB}
(Xie-Beni). The default \code{ALL} gives all indices. See \code{\link[fclust:Fclust.index]{fclust::Fclust.index()}}
}
\examples{
data(TS.sim)

fitsplines <- TPSfit(TS.sim, vars=c("Var1", "Var2", "Var3"), time="Time",
     ID="SubjectID", knots_time=c(0, 91, 182, 273, 365), n_fit_times=10)

ccoefs_2 <- cluster.coefs(fitsplines, k=2, seed=1234, RS=10)
ccoefs_3 <- cluster.coefs(fitsplines, k=2, seed=1234, RS=10)
ccoefs_4 <- cluster.coefs(fitsplines, k=2, seed=1234, RS=10)
ccoefs_5 <- cluster.coefs(fitsplines, k=2, seed=1234, RS=10)

# Compare clusters using all indices and custom names
comp.FKM.indices(list(ccoefs_2, ccoefs_3, ccoefs_4, ccoefs_5),
clusternames=c("k=2", "k=3", "k=4", "k=5"))

# Compare clusterings using a subset of the indices
comp.FKM.indices(list(ccoefs_2, ccoefs_3, ccoefs_4, ccoefs_5),
clusternames=c("k=2", "k=3", "k=4", "k=5"), index=c("SIL.F", "XB"))

}
\seealso{
Values for cluster validity indices are calculated using the
\pkg{fclust} package. See \code{\link[fclust:Fclust.index]{fclust::Fclust.index()}}
}
