% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMfunctions.R
\name{FKM.glm}
\alias{FKM.glm}
\title{Fit a generalized linear model (GLM) using clusters as predictors}
\usage{
FKM.glm(FKM_object, data, y, covariates, refclus = 1, family = "gaussian", ...)
}
\arguments{
\item{FKM_object}{An object of class '\code{FKM.TPS}' output from \code{\link[=cluster.fitted]{cluster.fitted()}}
or \code{\link[=cluster.coefs]{cluster.coefs()}}.}

\item{data}{A data frame with the same subjects used for spline-fitting and
clustering that includes an outcome variable of interest and optional
covariates.}

\item{y}{Name of the outcome variable (e.g. \code{y="Death"})}

\item{covariates}{A vector of covariates of interest to be included in the
model.}

\item{refclus}{Numeric identification of the cluster to be used as the
reference cluster. Default is cluster 1 (\code{refclus=1}). Use \code{refclus=0} to
identify the noise cluster as the reference cluster.}

\item{family}{A description of the error distribution and link function to be
used in the model.}

\item{...}{Additional arguments for the \code{\link[=glm]{glm()}} function.}
}
\value{
An object of class '\code{FKM.glm}' containing the following components:
\itemize{
\item \code{FKM_object } The inputted object of class '\code{FKM.TPS}'.
\item \code{model_data } A data frame containing the variables used in the model, including degree of cluster membership.
\item \code{formula } The formula used in the model.
\item \code{family } The family call used in the model.
\item \code{covariates} The covariates that were included.
\item \code{model_full } The GLM model using clusters as predictors and any additional
covariates of interest.
\item \code{model_noclusters } The GLM model using the covariates of interest but no
clusters.
\item \code{anova } ANOVA comparing the models with and without clusters as predictors.
\item \code{anova_pval } P-value for the ANOVA comparing the models with and without
clusters as predictors.
}
}
\description{
\code{FKM.glm()} fits a generalized linear model (GLM) using clusters output from
\code{\link[=cluster.fitted]{cluster.fitted()}} or \code{\link[=cluster.coefs]{cluster.coefs()}} as predictors, along with additional covariates.
}
\details{
\code{FKM.glm()} applies the \code{\link[=glm]{glm()}} function to fit a generalized linear model
using clusters as predictors. Clusters are obtained using \code{\link[=cluster.fitted]{cluster.fitted()}}
or \code{\link[=cluster.coefs]{cluster.coefs()}}, and the output object of class \code{FKM.TPS} is input into
the \code{FKM.glm()} function, along with a dataset containing the output variable
and additional covariates of interest. Clusters are included using the
"partial assignment" method that employs the degree of cluster membership for
each individual to account for uncertainty in the cluster assignment.
}
\examples{
data(TS.sim)

fitsplines <- TPSfit(TS.sim, vars=c("Var1", "Var2", "Var3"), time="Time",
     ID="SubjectID", knots_time=c(0, 91, 182, 273, 365), n_fit_times=10)

clusters1 <- cluster.fitted(fitsplines, k=3, m=1.3, seed=12345, RS=5, noise=TRUE)

model <- FKM.glm(clusters1, TS.sim, y="outcome", covariates=c("x1", "x2"),
family="binomial")
summary(model)
model$anova_pval

}
\seealso{
The glm function: \code{\link[=glm]{glm()}}
}
