% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{plot.FKM.glm}
\alias{plot.FKM.glm}
\title{Plot a GLM that uses clusters as predictors}
\usage{
\method{plot}{FKM.glm}(x, ...)
}
\arguments{
\item{x}{The output object of class '\code{FKM.glm}'.}

\item{...}{Additional arguments.}
}
\value{
A \code{plot.glm} plot.
}
\description{
Employs the \code{plot.glm} function to plot the generalized linear model with clusters and additional covariates as predictors.
}
