% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{plot.FKM.predicted}
\alias{plot.FKM.predicted}
\title{Plot an object of class '\code{FKM.predicted}'}
\usage{
\method{plot}{FKM.predicted}(
  x,
  center = TRUE,
  xmin,
  xmax,
  ntime = 100,
  lab_x,
  lab_y,
  bw = TRUE,
  title,
  title_size = 15,
  axis_label_size = 15,
  axis_title_size = 15,
  legend_label_size = 15,
  strip_label_size = 15,
  type = "raw_grid",
  ...
)
}
\arguments{
\item{x}{An object of class '\code{FKM.predicted}'}

\item{center}{Logical expression indicating whether trajectories are centered
on individual means.}

\item{xmin, xmax}{Optional minimum and maximum values to show on x-axis.}

\item{ntime}{Optional number of times to calculate fitted values for smoothed
plots.}

\item{lab_x, lab_y}{Optional labels for x- and y-axis.}

\item{bw}{Logical expression for black and white graphic.}

\item{title}{Optional title.}

\item{title_size}{Optional title size.}

\item{axis_label_size}{Optional size of axis labels.}

\item{axis_title_size}{Optional size for axis titles.}

\item{legend_label_size}{Optional size for legend.}

\item{strip_label_size}{Optional size for strip labels on graphics.}

\item{type}{Type of plot to produce. Options are \code{"raw"}, \code{"raw_grid"}, \code{"smooth"},
and \code{"smooth_grid"}.}

\item{...}{Additional arguments}
}
\value{
A plot of raw or smoothed trajectories for new data by predicted cluster assignment.
}
\description{
Plot an object of class '\code{FKM.predicted}'
}
\examples{
data(TS.sim)

fitsplines <- TPSfit(TS.sim, vars=c("Var1", "Var2", "Var3"), time="Time",
     ID="SubjectID", knots_time=c(0, 91, 182, 273, 365), n_fit_times=10)

clusters1 <- cluster.fitted(fitsplines, k=3, m=1.3, seed=12345, RS=5, noise=TRUE)

predicted_clusters <- predict(clusters1, TS.sim.new)
plot(predicted_clusters, label_size=10, type="raw_grid")
}
