% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{plot.TPSfit}
\alias{plot.TPSfit}
\title{Plot an object of class \code{TPSfit}.}
\usage{
\method{plot}{TPSfit}(
  x,
  center = TRUE,
  xmin,
  xmax,
  ntime = 100,
  lab_x,
  lab_y,
  bw = FALSE,
  title,
  title_size = 15,
  axis_label_size = 15,
  axis_title_size = 15,
  legend_label_size = 15,
  strip_label_size = 15,
  type = "raw",
  ...
)
}
\arguments{
\item{x}{Object of class \code{TPSfit}.}

\item{center}{Logical expression indicating whether to center data on individual trajectories.}

\item{xmin, xmax}{Optional minimum and maximum values to show on x-axis.}

\item{ntime}{Optional number of times to calculate fitted values for smoothed plots.}

\item{lab_x, lab_y}{Optional labels for x- and y-axis.}

\item{bw}{Logical expression for black and white graphic.}

\item{title}{Optional title.}

\item{title_size}{Optional title size.}

\item{axis_label_size}{Optional size of axis labels.}

\item{axis_title_size}{Optional size for axis titles.}

\item{legend_label_size}{Size of legend label, if applicable}

\item{strip_label_size}{Optional size for strip labels on graphics.}

\item{type}{Type of plot to produce. Options are "smooth", "raw", and "all".}

\item{...}{Additional arguments}
}
\value{
A plot of the raw or smoothed data.
}
\description{
Plots an object of class \code{TPSfit}, with option to plot either raw or smoothed data.
}
\examples{
library(tidyr); library(dplyr); library(mgcv); library(ggplot2)
data(TS.sim)

fitsplines <- TPSfit(TS.sim, vars=c("Var1", "Var2", "Var3"), time="Time",
     ID="SubjectID", knots_time=c(0, 91, 182, 273, 365), n_fit_times=10)
plot(fitsplines, type="smooth")

}
