% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterfunctions.R
\name{predict.FKM.TPS}
\alias{predict.FKM.TPS}
\title{Predict the cluster assignment for new subjects}
\usage{
\method{predict}{FKM.TPS}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class '\code{FKM.TPS}' of previously identified
clusters.}

\item{newdata}{A new data frame containing subjects with the same variables
previously used for clustering.}

\item{...}{Additional arguments.}
}
\value{
An object of class '\code{FKM.predicted}' containing the following components:
\itemize{
\item \code{predicted_U } A data frame containing the subjects with their degree of cluster
membership and modal cluster assignment.
\item \code{U } The degrees of cluster membership for each subject.
\item \code{Umax } The modal class assignment for each subject.
\item \code{wide_data } A data frame containing the values that were clustered on.
\item \code{IDmatch } A data frame matching the original subject ID variable to a new
consecutive subject ID.
\item \code{TPSdata } An object of class '\code{TPSfit}' containing the fitted splines for the
new subjects.
\item \code{FKM_TPS } The inputted object of class '\code{FKM.TPS}'.
\item \code{noise } Logical expression indicating whether a noise cluster is included for
outliers.
\item \code{k } The number of clusters.
\item \code{m } The fuzziness parameter.
}
}
\description{
After clustering a set of trajectory data, predict the cluster assignment for
new subjects.
}
\examples{
data(TS.sim)

fitsplines <- TPSfit(TS.sim, vars=c("Var1", "Var2", "Var3"), time="Time",
     ID="SubjectID", knots_time=c(0, 91, 182, 273, 365), n_fit_times=10)

clusters1 <- cluster.fitted(fitsplines, k=3, m=1.3, seed=12345, RS=5, noise=TRUE)

predicted_clusters <- predict(clusters1, TS.sim.new)
summary(predicted_clusters)

}
