% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPSClus-package.R
\docType{package}
\name{TPSClus-package}
\alias{TPSClus}
\alias{TPSClus-package}
\title{TPSClus: Clustering Multivariate Longitudinal Data using Tensor-Product Splines}
\description{
Functions for fitting tensor-product splines to multivariate longitudinal data, fuzzy clustering on the fitted values of splines or spline coefficients.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://silivingston.github.io/TPSClus/}
  \item \url{https://silivingston.github.io/}
}

}
\author{
\strong{Maintainer}: Sherry Livingston \email{sherry.i.livingston@gmail.com} (\href{https://orcid.org/0000-0003-2016-9957}{ORCID})

}
\keyword{internal}
