% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\name{TS.sim.new}
\alias{TS.sim.new}
\title{A data frame of simulated multivariate time series data for predictions}
\format{
A data frame with 325 observations of 7 variables:
\describe{
\item{\code{SubjectID}}{Identifier for each subject}
\item{\code{Time}}{Time variable ranging from 0 to 365.}
\item{\code{Var1}, \code{Var2}, \code{Var3}}{Variables to be smoothed and clustered on.}
\item{\code{x1}, \code{x2}}{Covariates for modeling the outcome.}
}
}
\description{
A data frame of simulated multivariate time series data for 30 individuals over one year.
}
