% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{TS.sim}
\alias{TS.sim}
\title{A data frame of simulated multivariate time series data}
\format{
A data frame with 1919 observations of 8 variables:
\describe{
\item{\code{SubjectID}}{Identifier for each subject}
\item{\code{Time}}{Time variable ranging from 0 to 365.}
\item{\code{Var1}, \code{Var2}, \code{Var3}}{Variables to be smoothed and clustered on.}
\item{\code{x1}, \code{x2}}{Covariates for modeling the outcome.}
\item{\code{outcome}}{Binary outcome variable.}
}
}
\description{
A data frame of simulated multivariate time series data for 150 individuals over one year.
}
